LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-form-control-expected.txt

Open popover      Close popover This button is where focus should land after traversing this popover

FAIL Focus navigation for form controls contained in a popover assert_equals: Focus should move forward from invoker to target1 expected Element node <input type="checkbox" id="target1"></input> but got Element node <body><div>
  <button popovertarget="popover" id="invoker...
